+++ src/main/java/SocialNetwork/SocialNetwork/repositories/PostRepository.java
@@ -3,14 +3,12 @@ package SocialNetwork.SocialNetwork.repositories;
 import SocialNetwork.SocialNetwork.domain.entities.Post;
 import SocialNetwork.SocialNetwork.domain.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post,Integer> {
+public interface PostRepository extends JpaRepository<Post,Long> {
     List<Post> findByUser(User user);
     List<Post> findByUserAndStatus(User user,Integer status);
     List<Post> findAllByStatus(Integer status);
